pointing-simulate: arXiv identifier added to BibTeX entry
Until now, the BibTeX entry (from ADS) for the paper on the pointing
pattern simulation script did not have an arXiv identifier (which is more
commonly used than the journals themselves!). But today, the paper was
published on arXiv, so we finally have it.

With this commit, the relevant arXiv entries have been added to the BibTeX
entry for the paper of this script.