astscript-rgb-image: added the make check test for this script
Until now, this script did not have a make check test. Also, the
documentation book was missing some hierarchy references.

With this commit, the make check test that correspond to this script has
been added. It takes the same image three times to generate a color image.
In addition to this, the hierarchy of the Book has been updated and now it
contains the missing references and links to the recent section related to
this script.