astscript-rgb-asinh: new name for the previous astscript-rgb-image
Until this commit, the name of the script for generating the RGB color
image using the asinh transformation was not informative enough.
'rgb-image' could be a lot of things. In order to be more explicit, the
'asinh' word should be included.

With this commit, the script has been renamed. Now it is:
"astscript-rgb-asinh". With this change, it is clear that it uses the asinh
transformations to change the pixel values and generate the color image.

In addition to this, a small subsection to the "Color image using asinh
transformation" section has been added. Here a small example on how to use
this script with the SDSS images is presented.