zeropoint: HDU of the input is used in all the commands
Until now, when we used the inputs for the 'astquery', the '--hdu' value
that the user gave was not used. When the inputs were in 1st HDU the script
worked propely (because this was the default for 'astquery'). However, when
the inputs were in the 0th HDU, the script crashed (because it could not
find the input). For more, see https://savannah.gnu.org/support/?110952.

With this commit, when the input image is given to 'astquery', the HDU
given by the user is also pased to 'astquery'. A minor clarification has
also been implemented in the book.

This bug was reported by 'danc' on the Savannah issue above.

This fixes bug #64852.