Book: new tutorial describing how to generate color images
Until this commit, there were some examples on how to generate color
images. However, a dedicated tutorial on color images generation was
missing.

With this commit, I have included a new tutorial that explains how to
create color images. It is based on the previous section about the aligning
of images to generate color images.  The tutorial is thought to have three
sections:

 1. Color channels in same pixel grid. This is something that was already
    present. But now it is the first subtsection.
 2. Color image using linear transformation. This section explain with some
    detail the use of `astconvertt' to generate color images. This is the
    new section added with this commit.
 3. Color image using asinh transformation. In this commit, this section
    has not been modified.