Book: subsection of rgb-asinh script added to the color images tutorial
Until this commit, the subsection of the tutorial for creating color images
using the astscript-rgb-asinh script was missing.

With this commit, this new section has been added. Right now it is at the
first stage. It contains information that can be improved in terms of
clarity and shortness. This will be done soon in next commits. In addition
to this, the tips that are shown on the command-line has been changed in
order to be more clear.