astscript-rgb-faint-gray: set gray background by default, new script name
Until this commit, the script that generated the color image with the gray
background has a very specific name containing the 'asinh' word to inform
that it uses that function internally. However, in order to be more general
and be able to include other functions in the future, it is good to remove
this word and be more general. In addition to that, the unique feature of
this script is its capacity of constructing the gray background, but this
was not the default option.

With this commit, necessary changes have been done with the goal of:
1. Construct the gray background image by default.
2. Change the name of the script to inform that the faint regions will be
shown in gray.