astscript-rgb-faint-gray: following Gnuastro standards for HDU arguments
Until this commit, the way HDUs were specified was not the standard by
Gnuastro. The origin of this is that the script was not originally part of
Gnuastro. However, in order to include this into Gnuastro, this issue has
to be fixed to follow the Gnuastro standards.

With this commit, the way HDUs are specified has been changed. Now it
follows the Gnuastro standards (i.e., multiple HDUs have to be provided,
one for each input image, or alternatively, use --globalhdu).