astscript-rgb-faint-gray: removing comma-separated arguments
Until this commit, there were some options of the script that took
comma-separated values. But this is not the standard way that Gnuastro
programs work. On the contrary, if it is necessary to specify an argument
for different inputs, the same argument has to be repeated in the
command-line invokation.

With this commit, some arguments (--minimum, --zeropoint, --weight) have
been changed in order to fix this issue. Now they follow the Gnuastro
standards.