Library (arithmetic.c): *clip-fill intermediate steps on threads
Until now, the part of the filled re-clipping operators that filled the
holes was done in series, not in parallel. For large and numerous datasets,
this could slow down the process significantly.

With this commit, the part where we find and fill the holes is done in
threads. Also, a limit of 1/50th the size of the full dataset was set for
the largest hole size to fill (we should later bring this out as a
parameter).