Book: improving the tutorial for rgb-faint-gray script by using J-PLUS data
Until this commit, this tutorial used SDSS data. However, we have J-PLUS
data that can be used for the step-by-step demostration.

With this commit, now J-PLUS images are used for explaining how to create
color images showing the low surface brightness structures. In addition to
that, other minor corrections have been done.