Book: edited the color tutorial, typo in rgb script fixed
Until now, Raul had edited the tutorial on the newly added RGB script with
a gray background to describe how the images in the paper are
produced. However, the demonstration commands were still too long and hard
to read, they didn't use an '=' between the option name and value (which is
the general convention in the manual), and the context behind some of
choices was not described.

With this commit, the issue of long commands has been fixed by using shell
variables to keep each filter's file name with its HDU, minimum value,
weight and zeropoint. Also, the option names and values now have an '=' and
the connectivity of the text has been improved.

In the process of testing the commands, I noticed that the '--minimum=AAA'
way to read the '--minimum' option would not work! The problem was a typo
in this scenario that has also been fixed with this commit.