Book: Changed the output directory to read codes easier
Until now, in the general tutorial's section where we show how to make
publication ready images, a seperate directory was created for saving the
outputs and the commands were run from outside of it. This resulted in
longer file paths for the output/input files of the commands that would
make it hard for a new user. Also, there were some typos in the comments of
some parts of the library.

With this commit, to keep the codes shorter and easier to read in the
tutorial, we first 'cd' into that directory and can use much
shorter/simpler commands. Fortheremore, the typos have been fixed.