color-faint-gray: new name for the script, --bias replaces --brightness
Until now, the script was called 'rgb-faint-gray', but RGB is only one
format of a color space and would limit us if we wanted to add other color
spaces like CMYK. Also, the option '--brightness' could be confusing
because this term has physical meaning in astronomy (although it is
regularly used in image processing!).

With this commit, the script is now called 'color-faint-gray' and the
'--brightness' option has been replaced with '--bias'.