Library (tableintern.c): message with bad column name fixed
When Table is given a column name that does not match the names of the
columns of its input it will recommend to use '--info' (as shown in the the
message in the P.S.). However, given the newly added Table options that
start with '--info' (for example '--info-num-cols' or '--info-num-rows'),
the '--info' suggestion is be to be ambiguous and Table will ask to be more
clear! This can be confusing for a new user and should be corrected.

With this commit, we are now using the full version of the option
('--information') in the error message.

This bug was reported by Sepideh Eskandarlou.

This fixes