astscript-psf-unite: correctly accounting for --innerhdu
Until now, when '--innerhdu' was used with this script to specify a HDU
other than '--innerhdu=1', the script would not use the user-specified HDU.

With this commit, the source of this problem has been found and fixed. It
was due to the 'astwarp' command where we were not passing the value give
to this option.

This bug was reported by Ignacio Ruiz Cejudo.

This fixes bug #65005.