color-faint-gray: bug corrected in the asinh transformation
Until this commit, the asinh transformation was not properly done! This was
noticed by Mohammad when changing the --qbright and --stretch parameters,
he noticed that only the ration between them was important. However, these
two parameters controls the different part of the transformation. The asinh
transformation was (wrong):

f(x) = asinh (qbright * x / stretch) / qbright

With this commit, the transformation has been set to the correct one:

f(x) = asinh (qbright *  x * stretch ) / qbright

With this transformation, the idea is to set --qbright to very low values.
This will make a linear transformation whose slope is then controled by
--stretch to show the fainter regions. Then change again --qbright to show
the bright parts.