color-faint-gray: improving the selection of color, black, and gray regions
Until this commit, the way the different regions (color, black, and gray)
were defined was too dificult and strange for the user.

With this commit, a single image has been set for defining the three
regions. From this threshold image, two values are used to separate the
three different regions.

By default, the threshold image is the asinh-transformed image with pixel
values from 0 to 100. From this image, '--colorval' is the lowest value to
be shown in color. In the same way, '--grayval' is the highest value to be
shown in gray. Consequently, the pixel values between this two parameters
are shown in black. These two values are estimated by the script but the
user can select other values to modify the regions.

If a fourth image is provided, then it is used for defining the different
regions as explained above.  We assume that the threshold image has higher
pixel values for brighter pixels. If this is not the case, special
attention to the --colorval and --grayval parameters is required. One
example of this situation is the use of a magnitude image for the threshold
image (fourth image). In this case, higher pixel values mean lower
brightness. As a consquence, the parameters are inverted.