astscript-fits-view: not accounting for different HDUs bug fixed
Until this commit, this script was not able to account for different HDUs.
It only considered the last HDU provided from the command line.

With this commit, this bug has been fixed. Now when specifying different
HDU numbers or names, they are properly accounted for.

This fixes bug #64825, reported by Teet Kuutma:
https://savannah.gnu.org/bugs/?64825