color-faint-gray: new option --segment to define the color regions
Until this commit, the way of separation the color, black, and gray regions
were internally computed. These regions could be changed with two
parameters, but it was always done by the script internally. However, in
some situations it is useful to define these regions in a quantitative way.

With this commit, a new option, '--segment' has been added. With this
option it is possible to provide a segmentation image as an input that will
be used for defining the different color regions.  Pixels equal to 2 in
this image will be shown in color, pixels equal to 1 in pure black, and
pixels equal to 0 will be shown in gray. Other minor modifications have
been also done.