Library (arithmetic.h): new operator to rotate a point
Until now, if you wanted to rotate a set of points within a table, you had
to do the mathematics yourself in a complex column arithmetic operation!

With this commit, to simplify the job, a new 'rotate-coord' has been added
to the general Arithmetic operators (so it can be used in both the
Arithmetic and Table programs).