Libraries (Makefile.am): added --libtool option to the TEMPLATE
Until now, there was not the --libtool option in the TEMPLATE instance when
you wanted to add the name of a new program in some lines of the
Makefile.am file in the doc directory. So, there was this question of if it
was an essential option for the new program. Also, there were several typos
in the book and libraries.

With this commit, This option has been added to the TEMPLATE instance.
Also, these typos within the book and comments of the code have been
corrected.

This fixes bug #65050.