Library (fits.h): Columns of empty FITS tables reversed to fit input
Until now, when an input FITS table contained 0 rows (was empty of data),
the columns (their metadata) was read in reverse! This happened because we
would simply add an empty dataset with the given input's metadata, but
using the 'gal_list_data_add_alloc' function which is first-in-first-out
(hense reversing the order!).

With this commit, after all the columns are read, the allocated table is
reversed to be the same as the input.

This fixes bug #65055.