All programs: ui.c function name correctly describes purpose
Until now, the 'ui_read_check_only_options' function of all the programs
included a '_read_' component. However, in this function, we only check the
option values, we don't read them. So, this function name does not
accurately reflect its purpose and this word is redundant.

With this commit, the word '_read_' component has been removed from the
function name to make it more descriptive and accurate.