Book: Tutorial of color-faint-gray broken into three sections
Until now, the section describing 'astscript-color-faint-gray' in the
tutorial was very long, making it hard to follow for a new reader. Also,
when printing the default and used values to the '--qbright', '--stretch',
'--colorval' and '--grayval' options, they were printed in one long line
(making it hard to read or parse automatically).

With this commit, after a review of the tutorial, it was broken into three
sections (the original one, added with another for manually selecting color
ranges and weights). Also, in the printed output of the script, the default
and used values of the options above are printed as a table.