Fits: SPACE printed between outputs of --skycoverage
Until now, when the Fits program was run with the '--skycoverage' option
without '--quiet', if the precision of one dimention was high (as in data
cubes with angstroms in the third dimension), it would connect the numbers
for the minimum and maximum value along that dimension.

With this commit, the issue has been solved by adding a SPACE character
between the formatted print that generated them.

This bug was found with the help of Rahna Payyasseri Thanduparackal.

This fixes bug #65084.