astscript-zeropoint: RNAAS publication added to --cite
Until now, the only recommendation for '--cite' of the zero point script
was Gnuastro's main paper of 2015. However, a dedicated research note has
been published recently on this script and users should primarily cite that
when using this script.

With this commit, the ADS BibTeX entry for the RNAAS paper has been added
to the output of '--cite' in this script. On the other hand, the citation
of the pointing simulate script had one extra word which has been
corrected.