ConvertType and Crop: the 0-th extension includes all metadata
Until now, ConvertType didn't write any metadata in its output FITS files
and Crop would write its metadata in the HDU that contained data. This was
contrary to the new idea that Gnuastro's data should only have WCS metadata
(which is necessary to locate it on the sky). Other metadata should be
written in the 0-th.

With this commit, both issues have been addressed.