Library (fits.h): encoded datasum string
Until now, Gnuastro's 'fits.h' library would only return the raw (integer)
datasum value. But in some scenarios, having the fixed-length, ASCII
encoded string which is available for the 'CHECKSUM' keyword in FITS is
more useful.

With this commit a function has been added to the library for this and the
new '--datasum-encoded' option has been added to the Fits program.