Book: example of radial profile script with MakeProfile's customtable
Until now, the example within the description of MakeProfile's custom table
only had an abstract (hand-written!) radial profile. In practice, this
profile will most likely be made by a script like
'astscript-radial-profile'. But there are some intricacies with using that
script (in particular the oversampling).

With this commit, to make it clear, a full script has been added to the
description, allowing the readers to try and see the effect themselves.

This issue was reported by Sepideh Eskandarlou.

This fixes bug #65106.