Library (statistics.h): MAD corrected for unsigned types
Until now, in the low-level MAD calculation function
('statistics_mad_in_sorted_no_blank'), I had forgot to include a 'break' in
the 'case' statement for unsigned types! As a result, when the inputs where
unsigned, it would always go to the 'default' scenario, resulting in a zero
MAD value! This would later cause problems in the newly added MAD-clipping
functions (where they would complain about the size).

With this commit, the 'break's have been added so each unsigned type is
converted to its respective larger signed type properly.

This issue was found with the help of Alberto Moreno Signes.