Library (statistics.h): clipping accounts for a spread of zero
Until now, when the spread (standard deviation in sigma-clipping or MAD in
MAD-clipping) became zero, the clipping function would fall into an
infinite loop, resulting in a final error about the size of the inputs!

With this commit, when checking if we have passed the tolerance, we also
check if the spread has become zero or not. If so, then the clipping stops.

While investigating the cause of the problem above (and checking if there
is any memory leaks), I noticed that when writing the value of linked-list
options, we weren't allocating the space for the value. This has also been
corrected in 'options.c' in this commit.