Library (arithmetic.h): correct name for *-clip-fill-* error messages
Until now, when reporting an error on any of the '*-clip-fill-*' operators
(for example 'madclip-fill-mean'), the library would say that the operator
name is '*-clip-dilate-*'! This was the original name of these operators
(before filling holes was included resuling in the name change).

With this commit, their names have been corrected.