MakeCatalog: informative error when labeled image is blank
Until now, when the labeled image was all blank, MakeCatalog would complain
with a cryptic error that was hard to unerstand (it crashed!).

With this commit, a check has been added that in such cases, MakeCatalog
aborts with the same error message as the case when there are no non-zero
labels.

This bug was reported by Sepideh Eskandarlou.

This fixes bug #65141.