MakeCatalog: error when --frac-max not given for relevant columns
Until now, when the user asked for any of the '--frac-max*-*' columns, but
forgot to give fractions to '--frac-max', MakeCatalog would continue to the
end! Furthermore, when asked for '--frac-max2-sum', the output column was
called 'FRAC_MAX1_SUM'.

With this commit, checks have been added for the following scenarios:

 - If '--frac-max' is not given (when those measurements are requested),
   MakeCatalog aborts with an error, guiding the user on how to progress.

 - If one value is given to '--frac-max', but the user asks for a second
   fraction's measurements, it aborts and guides on how to progress.

Also, the problem with the wrong column name has been fixed. Furthermore,
thanks to Raul Infante-Sainz, I noticed that we had not included the fix to
bug #64825 in the 'NEWS' file; so that was also added.

This bug was reported by Helena Domínguez Sánchez.

This fixes bug #65149.