Library (txt.h): new func. to read words file as list of strings
Until now, there was no easy way to pass a very long list of arguments to
the Fits program: the shell would abort with an "Argument list too long"
error and abort!

With this commit, a new library function has been added to Gnuastro's
plain-text library to read a list of "words" within a file and prepare them
as a list. Using that, the Fits program now has a new '--infilelist' option
to allow the users to give the list of input files within a file.