Book: rectified functions including the 'hdu_option_name' option
Until now, there were two names for one parameter: 'hdu_option_name'
and 'ext_option_name'. Besides, the the recent changes in adding
'hdu_option_name' option were not yet fully added in some functions
and example of the book.

With this commit, the 'ext_option_name' has been changed to
'hdu_option_name'. Also, this option has been added to some of the
functions and examples of the book that had changed in the libraries
of Gnuastro but had not changed in the book. Plus, the definition of
this option moved to 'gal_array_read' instead of 'gal_fits_hdu_open'
where the user faces this option for the first time.