Fits: renamed the newly added option to --arguments
Until now, the newly added option to specify the list of input files was
called '--infilelist'. But this option is going to be necessary later in
other programs too: Arithmetic in particular. And in that case, the
arguments are not necessarily file names, but can also contain operators.

With this commit, the option name has been renamed to '--arguments' to be
usable in Arithmetic also later.