Arithmetic: new --arguments to allow very long argument lists
Until now, if the list of arguments given to Arithmetic was too long, the
shell would abort with an "Argument list is too long" error. In the last
few commits an '--argument' option was added to the Fits program. This can
be useful for Arithmetic also.

With this commit, this option has been added to Arithmetic