Statistics: new --checkskynointerp to avoid interpolation and smoothing
Until now, the only way to access the good tile values for sky measurements
of Statistics was the '--checksky' option. However, this would also do
interpolation and smoothing (which could take a significant amount of time.

With this commit, to avoid wasting time in such cases, this new option has
been added which will abort Statistics once the good tiles have been
identified.