astscript-ds9-region: check for existence of input file
Until now, when the input file given to this script did not exist, it would
still take it and prepare the output (before checking if it actually
exists!). As a result, even when it crashed and made an empty output, the
users would see an output, confusing users.

With this commit, before registering the input file name as an input, this
script will check if it actually exists and abort if it does not.

Also, if more than one input is given, the script will abort (informing the
user that only one input should be given). To avoid confussion within the
script, the variable name 'inputs' was changed to 'input'. This component
was added by Mohammad while reviewing Thorsten's commit.

This bug was reported by Sepideh Eskdandarlou.

This fixes bug #61476.