Build system: dependence of programs on library within Make
Until now, when a change was made to the library's source, the programs
would not be re-built (assuming they were already built). The problem was
that Gnuastro's library was only passed as a linking library with
'-lgnuastro'; therefore Automake would just pass it to 'ld'; not check it
as a prerequisite.

With this commit, the direct path of 'libgnuastro.la' is used:
$(top_builddir)/lib/libgnuastro.la. Therefore the makefile is able to add a
real dependency and Libtool will worry about the linking
details.

Unfortunately this requires more CPU cycles during development as every
binary has to be relinked if a change is made in the library. To fix this,
we have the '--enable-*' options at configure time (for example
'--enable-noisechisel'; available in all programs) that will only build the
requested programs during developement.