Adding the color-faint-gray arXiv references
Until this commit, we did not have a formal reference for the paper
presenting the 'astscript-color-faint-gray' script.  However, this paper
has been accepted in the RNAAS and submitted to arXiv.

With this commit, I have included this reference into the script, as well
as its arXiv URL into the Book. In addition to this, I modified a bit the
BibTeX text appearing on the 'astscript-zeropoint' script to follow the
same standards.