Library (cosmology.h): slower/robust gsl integration when fast fails
Until now, when CosmicCalculator was given an input redshift that is higher
than 100, GSL would produce a code-dump. This was due to the choice of
'gsl_integration_qng' for the numerical intergration (because it is fast).

With this commit, after 'gsl_integration_qng' is finished we check whether
it was successful or not. If it failed, the slower, but more robust
'gsl_integration_qag' is used. In order to use the 'error' function to
report the failure, the variable that was previously named 'error' was
renamed to 'gslerr'.