Book: unified format for citing arXiv papers
Until now, within the book we used multiple ways to cite arXiv papers
(sometimes printing the arXiv ID, sometimes the year and sometimes the full
author+year).

With this commit, a single format is used for all such links (only the year
is a link (as in published astronomical papers).