astscript-psf-stamp: default tmpdir includes location, not files
Until now, when this installed script was run, it would create a fixed
temporary directory (independent of the input), and the temporary files
would get an extra identifier for the input coordinates. This had two
problems:

 - When the user gave a desired temporary directory and wanted to keep it,
   they would still need to track the position of their desired object.

 - A new run of this script would put temporary files in the same
   directory; making it confusing and potentially buggy (for the user).

 - It is not the way other installed scripts work.

With this commit, the temporary directory now contains the position of the
desired cut, and the contents of a certain step always have the same
name.

Also, thanks to the reminder by Raul Infante-Sainz, we noticed that the
zero point paper was not actually mentioned at all in the book! So this
commit also includes several citations to that paper in the relevant parts.