astscript-radial-profile: BibTeX entry from ADS added
Until this now, we did not have a formal reference for the paper presenting
the 'astscript-radial-profile' script.  However, this paper has been
accepted in the RNAAS, submitted to arXiv and indexed in ADS.

With this commit, I have included this reference into the script (as the
output of '--cite'), as well as its arXiv URL into the Book. In addition to
this, I corrected a using example of Warp that was missing one option.