Library (eps.c): Clang literal-conversion warning fixed
Until now, when compiling Gnuastro, the following warning was printed by
Clang on macOS: "eps.c:879:29: warning: implicit conversion from 'double'
to 'long' changes value from 0.1 to 0". The warning was correct (and the
written value would always be zero, causing a human-reading confusion when
reading the code.

With this commit, '0' is used instead of the literal '0.1' to fix the
problem.

This issue was reported after a check by Raul Infante-Sainz and found
thanks to Thorsten Alteholz.

This fixes bug #52674.