NEWS: entry for change of previous commit and new publications
Until now, the NEWS file was not updated due to the change that was
introduced in the 'astscript-psf-stamp' script of the previous
commit. Also, there was no mention of new Gnuastro's publications.

With this commit, the previous commit's change has been described in the
'NEWS' file and the a new section has been added to list the
Gnuastro-related publications.