CosmicCalculator: allow curved FLRW models for radial distances
Until now, the two files 'bin/cosmiccal/ui.c' and 'lib/cosmology.c' would
prevent the calculation of cosmological distances for non-flat FLRW
models. If the aim of astcosmiccal is to be limited strictly to LambdaCDM,
then there should be much stricter limits on all the FLRW local
cosmological parameters, not just their sum.

With this commit, the error message in 'bin/cosmiccal/ui.c' is modified to
instead report the current value of the curvature density parameter, and
the message in 'lib/cosmology.c' is modified to remind the user that the
model is non-flat. In both cases, the error message does not abort the
program. To let the user avoid the library warning message, a new 'quiet'
argument has been added to all the affected 'gal_cosmology_*' functions.

Note that this change is *only* valid for the radial comoving distance,
i.e. the '--properdistance' at the current epoch. The angular diameter
distance is wrong for the non-flat cases, because it is not coded in
astcosmiccalc. Currently, cosmdist [1] can be used for more generic FLRW
distance/age/redshift calculations.

[1] https://codeberg.org/boud/cosmdist