CosmicCalculator: --angluardimdist changed to --angulardiamdist
Until now, the option for calculating the special case of the flat-space
angular diameter distance, i.e. the radial comoving distance divided by
(1+z), had the string "dim" (possibly alluding to "dimension"?)  rather
than "diam" to indicate "diameter".

With this commit, 'diam' is used instead of "dim" respectively to avoid
this confusion. The 'NEWS' file now includes this as a "Changed feature"
for people who may confront an error after updating to a new version.