astscript-color-faint-gray: fixed bug with weights of differing type
Until now, the script crashed when the user provided weights in different
data type formats. For example, if the user provides '-w1 -w1 -w1.5', two
of them are integers, while one is floating point. One of the Arithmetic
calls within the script would crash; complaining about different data
types.

With this commit, this bug has been fixed by forcing to consider the
weights as floating point numbers ('f32') in the step that they are
used. This problem was reported by Sepideh Eskandarlou.