Book: list of Gnuastro authors taken to third PDF page
Until now, the list of Gnuastro authors was printed on the second page
(where the copyright notice was). This was fine until the previous commit
by a new committer (Thorsten) when the list of authors was shorter than the
space above the copyright notice. However, when the previous commit by
Thorsten, that list became longer than the available space and the
copyright notice was broken in two pages.

With this commit, the list of authors is now printed on a third page and
now has space to grow as much as it needs :-).