NEWS: added bug fix entry of previous commit
Until now, we had forgot to mention the bug fix of the previous commit in
the NEWS file.

With this commit, a note has been added for this. Also the sanity check at
the start of the more complex cosmology library functions was moved after
the definitions of variables (as is necessary according to Gnuastro's
coding conventions).