astscript-color-faint-gray: added --markoptions to draw over output
Until now, it was not possible to easily draw markers on the final output
image generated by this script: the user would have to manually go into the
source of this script and add the necessary options.

With this commit, a new option '--markoptions' has been added. This option
allows to pass the necessary marker options to ConvertType in order to draw
the markers on the final output image.