Library (options.h): --quiet affects un-given --minmapsize warning
Until now, when any of the programs couldn't find the common '--minmapsize'
option, but were givn the'--quiet' option, they would still print the
warning about an un-given '--minmapsize'. This is a bug because the warning
message suggests using '--quiet' so it is no longer printed.

With this commit, a check of the value to '--quiet' has been added before
this warning is printed: so if it is given, it the warning is not
printed.

Also, while trying to track down the problem, I noticed that some of the
lines of 'bin/cosmiccal/ui.c' were a couple of characters larger than the
standard 75 character limit of Gnuastro's coding conventions. They have
also been fixed with this commit.

This bug was reported by Boud Roukema.

This fixes bug #65194.