NEWS: corrected formats of the bugs fixed section
Until now, the "Bugs fixed" section of the NEWS file for the current
version of Gnuastro didn't highlight the bug number (as we had done in
previous versions).

With this commit, that has been fixed by adjusting the subsequent lines of
each bug report to start after the bug identifier.