Book: spell check on additions since version 0.21
Until now, we hadn't ran a spell-check on the newly added sections of the
book since version 0.21. Also, while doing a test build on Conda, I noticed
that the release checklist for conda is missing a pull of the latest
changes.

With this commit, a spell-check was done on the newly added parts of the
book and the necessary commands in the conda release checklist have been
added.