Book: corrected outputs in STD vs error section
Until now, there was an ASCII histogram to show a normal distribution as an
example in the Standard deviation vs error section. However, the reported
statistics (values and histogram) from the commands (integers) did not
match the floating point outputs in the book. This happened because this
section was written before the Poisson distribution returned integers.

With this commit, the output of statistics has been replaced to more
precisely reflect the integer pixel statistics that the user sees after
running the command.