Book: more elaboration on the upper limit surface brightness
Until now, the "Quantifying measurement limits" section of the book
contained the "Upper limit magnitude of image", while the description was
about the upper limit surface brightness of an image! The description of
that section was also very minimal.

With this commit, the section has been renamed to "Upper limit surface
brightness of image" and an elaborate description has been added on how it
is derived.