Release: routine post-release operations
Until now, in the previous commit, Gnuastro 0.22 was released. To be
prepared for all subsequent work, some routine steps are necesseary.

With this commit, all the routine steps are taken based on the release
checklist, and necessary additions/modifications (based on the 0.22
release) to the checklist have also been made.