Clarify that arcsectandist is in physical kpc; other copyedits
Until now, the '--help' output and Info documentation for the
'--arcsectandist' option did not state whether the tangential distance was
calculated in comoving or physical distance units.

While it is natural to expect physical units for such a small angle,
someone might guess that multiplying by 3600 gives the tangential distance
for an arc subtending one degree. However, at one degree, the comoving
separation for redshifts from 1 to 10 is about 100 Mpc give or take 50% or
so, i.e. this is the cosmic web scale, where comoving units usually make
more sense than physical units.

With this commit, to avoid people not being sure what the intended unit is,
the word "physical" is added in the description to clarify that these are
the units of "ordinary" physics as opposed to comoving FLRW distance units.

Some other copyediting is also done in this commit, including a comment for
the proper distance that it is in comoving units (by definition). For
distances such as the angular diameter distance or the luminosity distance,
it would be meaningless to describe the units as either physical or
comoving, since these do not represent point-A-to-point-B distances in the
sense of the FLRW model.

Disclaimer: I have not checked the volume calculation.