Installation: system and user config files in etc/gnuastro/ subdir
Until now, Gnuastro would install the configureation files of the programs
in 'PREFIX/etc'. But this directory is used by many programs and can get
crowded and hard to manage.

With this commit, the system configuration files are installed in
'PREFIX/etc/gnuastro/' and the user configuration files are searched for in
the '$HOME/.local/etc/gnuastro/' directory.

This was suggested by Phil Wyett as bug #65249. In the 'NEWS' file, I am
not marking it as a bug, but treating it more like a "task" and have listed
it as "Changed features".