Book: edits to --arcsectandist of CosmicCalculator
Until now, the recently updated discussion for '--arcsectandist' option
could be confusing: but it's in physical units, not comoving units. Also,
the way to get comoving results was not given.

With this commit, a minimal edit has been made to the description of this
option, as well as an extra edit of the angular diameter distance.