astscript-fits-view: new --globalhdu like Arithmetic or ConvertType
Until now, when multiple inputs were given to this script, but only a
single HDU was given, the script would not check and call DS9 with an empty
'[]' (the way DS9 knows which HDU to open). As a result, DS9 would crash!

With this commit, this script now has '--globalhdu' (or '-g') with an
identical behavior to Arithmetic or ConvertType (that also take multiple
files). A check (with an elaborate description) has also been to make sure
that if '-h' (or '--hdu') is given, it is at least the same number as the
number of inputs. As a result of this change, it was necessary to change
the short format of the '--ds9geometry' option to '-G'.

This fixes bug #65267.