Library (pointer.h): solution for HPCs added to error message
Until now, the error message that was printed when enough memory was
available (but none could be allocated) only described the problem, without
any potential solutions.

With this commit, after finding the cause and solution of the problem with
the help of Juan Castillo Ramírez, the description has been modified to
propose solutions for a user that confronts it.