Book: differentiating MEDSTD from Standard Deviation in Masked Images
Until now, the tutorial has provided a clear explanation and examples for
estimating the image surface brightness limit using two methods. The first
method involves calculating the standard deviation of a masked image, while
the second method utilizes the MEDSTD keyword stored within the SKY_STD
HDU. While the tutorial mentions that the differences between MEDSTD and
the standard deviation of the masked image are generally small, it does not
clarify to the reader which is more precise when there is a difference.

With this commit, a box has been added in the text to highlight how MEDSTD
is not only easier to use, but is more robust to outliers; and therefore,
it is the one that should be prefered. Some minor corrections have been
made in the text.